backend/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
pydantic==2.5.3
pydantic-settings==2.1.0
python-dotenv==1.0.0
requests==2.31.0
beautifulsoup4==4.12.2
celery==5.3.4
redis==5.0.1
scikit-learn==1.3.2
numpy==1.26.3
pandas==2.1.3
cors==1.0.1
python-multipart==0.0.6
